Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-decouple==3.8
pydantic==2.10.6
pydantic-settings==2.7.1
openai>=1.0.0
anthropic>=0.7.0
google-genai
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.24.3
pandas>=1.5.0
python-dotenv>=1.0.0
transformers>=4.36.2
torch>=2.1.2
pdfplumber>=0.10.3
spacy>=3.7.2
prometheus-fastapi-instrumentator==6.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
stripe==7.9.0
boto3>=1.34.0
google-cloud-storage>=2.14.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
sentry-sdk[fastapi]>=1.39.0
celery>=5.3.0
# Vector Database Libraries
pinecone-client>=3.1.0
weaviate-client>=3.25.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
